requirements.txt
annotated-types==0.7.0
anyio==4.11.0
Authlib==1.6.5
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
deepmultilingualpunctuation==1.0.1
defusedxml==0.7.1
deprecation==2.1.0
filelock==3.19.1
fsspec==2025.9.0
grpcio==1.75.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
packaging==25.0
protobuf==6.32.1
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
neo4j==5.26.0
python-dotenv==1.0.1
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
sniffio==1.3.1
sympy==1.14.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.57.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
validators==0.35.0
weaviate-client==4.17.0
youtube-transcript-api==1.2.2
